Lesson-8: fix routes for entities
