+++ Project/public/queue/queue.js
@@ -1,7 +1,7 @@
 const socket = new WebSocket('ws://localhost:8080');
 
 async function getSelectedResolution() {
-  const response = await fetch('/queue/patient/resolution', {
+  const response = await fetch('/resolution', {
     method: 'GET',
     headers: {
       'Content-Type': 'application/json;charset=utf-8',
@@ -14,7 +14,7 @@ async function getSelectedResolution() {
 }
 
 async function getCurrent() {
-  const response = await fetch('/queue/patient/current');
+  const response = await fetch('/queue/current');
   const result = await response.json();
   document.getElementById('currentNumber').innerHTML = result.name;
   await getSelectedResolution();
