--- Project/routes/main.js
@@ -1,9 +1,0 @@
-import express from 'express';
-
-const router = express();
-
-router.get('/', (req, res) => {
-  res.redirect('/login');
-});
-
-export default router;
