+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "build": "webpack",
     "start": "node --experimental-specifier-resolution=node Project/index.js",
-    "lint": "eslint ./Project/api/**/*.js ./Project/routes/*.js ./Project/*.js --fix",
+    "lint": "eslint ./Project/api/**/*.js  ./Project/*.js --fix",
     "test": "jest --runInBand"
   },
   "jest": {
